+++ changelog.md
@@ -1,9 +1,12 @@
 # FluxLoader v0.2.1
 - The system for subscribing classes as listeners has been fixed. Now it is not the class itself that is signed, but its instance.
+- New event added `onServerShutdown`
+
 # FluxLoader v0.2.0
 - Redesigned plugin loading system
 - Changed plugin creation template
-- Added metadata logic for the plugin
+- Added metadata logic for the 
+
 # FluxLoader v0.1.0
 - Beta version of the implementation of the basic functionality of the plugin loader.
 - An event system has been created.